backend/requirements.txt
newspaper3k
uvicorn
python-dotenv
alpaca-py
openai
requests